Add some missing readonly checks:
* Do not allow working on Special:EditWatchlist
* Do not reset all notification markers
* Do not delete expired restrictions

+++ includes/Title.php
@@ -2777,6 +2777,10 @@ class Title {
 	 * Purge expired restrictions from the page_restrictions table
 	 */
 	static function purgeExpiredRestrictions() {
+		if ( wfReadOnly() ) {
+			return;
+		}
+
 		$dbw = wfGetDB( DB_MASTER );
 		$dbw->delete(
 			'page_restrictions',
